Raise MissingParameterError when the input file is not given.
When the input file is not given to Decompiler.start_decompilation() or
Fileinfo.start_analysis(), we should raise MissingParameterError. The
retdec.com API would raise an exception, anyway. What is even more important,
Decompiler.start_decompilation() would not even send the request when input
file was not given (it would fail on an internal exception that tries to access
the name property on None).