Add support for decompilation of files in archives.
RetDec 2.2 added support for decompilation of files in archives (ar format).
Archives are statically linked libraries, commonly ending with .a or .lib. To
decompile a file from an archive, use the bin mode and pass either the index of
the file in the archive (ar_index) or its name (ar_name).