Fix the use of environment variable RETDEC_API_URL to override the default API URL.
Previously, the URL was not overridden even when given. The reason was that our
tools (e.g. decompiler and fileinfo) set a default value of the -u/--api-url
parameter, which overrode the RETDEC_API_URL environment variable. Thus,
RETDEC_API_URL was never applied when using the tools.

The solution is to set the default API URL to None in the tools and let the
Service base class handle the environment variables.